+++ R/bca.R
@@ -8,6 +8,7 @@
 #' @param tt Mandatory. A (0,1)-matrix or a boolean matrix. The number of columns must match the number of elements (values) of the frame of discernment \eqn{\Theta}. Each row is a subset of \eqn{\Theta}. The last row is the frame \eqn{\Theta}, represented by a vector of 1's.
 #' @param ssnames A list of subsets names which will be obtained from the column names of the tt matrix.
 #' @param m A numeric vector of length equal to the number of rows of the matrix  \code{tt}. Values of \code{m} must lie in the interval \code{(0,1]} and must add to one. The mass \code{m(k)} represents the chance value allotted to the proposition represented by the row \code{k} of the matrix \code{tt}.
+#' @param qq  Commonality functions from the frame of discernment to \eqn{[0,1]}
 #' @param cnames A character vector containing the names of the elements of the frame of discernment \eqn{\Theta}. The length must be equal to the number of elements of \eqn{\Theta}. The names are first searched in the \code{valuenames} parameter. If NULL, column names of the matrix \code{tt} are taken if present. Otherwise, names are generated.
 #' @param con The measure of conflict can be provided. 0 by default. 
 #' @param idvar The number given to the variable. A number is necessary to manage relations between variables  and make computations on a graph. 0 if omitted. 
@@ -17,6 +18,7 @@
 #' @param inforel Not used here. Defined within function \code{\link{bcaRel}}.
 #' @return y An object of class \code{bcaspec} called a bca for "basic chance assignment": \itemize{
 #'   \item tt  The table of focal elements. Rownames of the matrix of focal elements are generated from the column names of the elements of the frame. See \code{\link{nameRows}} for details.
+#'   \item qq  Commonality functions from the frame of discernment to \eqn{[0,1]}
 #'   \item spec  A two column matrix. First column contains numbers given to the subsets, 1 to  \code{nrow(tt)}. Second column contains the mass values of the subsets. 
 #'   \item con  The measure of conflict.
 #'   \item infovar  The number of the variable and the size of the frame of discernment.
@@ -48,7 +50,7 @@
 #' \item Shafer, G., (1976). A Mathematical Theory of Evidence. Princeton University Press, Princeton, New Jersey, p. 38: Basic probability assignment.
 #' \item Guan, J. W. and Bell, D. A., (1991). Evidence Theory and its Applications. Elsevier Science Publishing company inc., New York, N.Y., p. 29: Mass functions and belief functions 
 #' }
-bca<-function(tt = NULL, m, cnames = NULL, con = NULL, ssnames = NULL, idvar = NULL, infovar = NULL, varnames = NULL, valuenames = NULL, inforel=NULL) {
+bca<-function(tt = NULL, m, qq = NULL, cnames = NULL, con = NULL, ssnames = NULL, idvar = NULL, infovar = NULL, varnames = NULL, valuenames = NULL, inforel=NULL) {
   #
   # Local variables: ztable, zdup, zframe, znames
   # Functions calls: nameRows, DoSSnames
@@ -134,12 +136,18 @@ bca<-function(tt = NULL, m, cnames = NULL, con = NULL, ssnames = NULL, idvar = N
     else {
       znames <- ssnames
     }
+    
+    # 7.1 build qq
+    if (is.null(qq) == TRUE ) {
+      qq <- commonality(tt,m)
+    }
+    
     #
     # 8. Construction of the result
     #
     rownames(tt) <- nameRows(tt)
     #
-    y<-list(con = con, tt = tt, spec = spec , infovar = infovar, varnames = varnames, valuenames = valuenames, ssnames = znames, inforel = inforel) 
+    y<-list(con = con, tt = tt, qq=qq, spec = spec , infovar = infovar, varnames = varnames, valuenames = valuenames, ssnames = znames, inforel = inforel) 
     # end test
     #
     class(y) <- append(class(y), "bcaspec")